Fixing auth check for jwt header parsing. No longer looking for a token type in the request. Fixing initialization reference to vertx and config in the TransferTaskDatabaseVerticle. The variable is not always set when instantiated by reflection. Using the getter instead. Adding common options to db vertical setup. Fixing type definitions for hsqldb table columsn to properly cast long as a BIGINT. TransferTask DB tests now passing. Refactored transfertasks `total_skipped` column name to `total_skipped_files` so it properly converted to the variable name upon camel/snake case conversion when marshalling json.

+++ agave-transfers/transfers-api/src/test/java/org/agaveplatform/service/transfers/database/TransferTaskDatabaseVerticleTest.java
@@ -112,7 +112,7 @@ public class TransferTaskDatabaseVerticleTest {
 
                         JsonObject updatedJsonObject = new JsonObject()
                                 .put("attempts", 1) // attempts
-                                .put("bytesTransferred", Math.pow(2, 19)) // getBytesTransferred())
+                                .put("bytesTransferred", (long)Math.pow(2, 19)) // getBytesTransferred())
 //                                .add("created should not update") // getCreated())
 //                                .add("dest should not update") // getDest())
                                 .put("endTime", Instant.now()) // getEndTime())
@@ -123,13 +123,13 @@ public class TransferTaskDatabaseVerticleTest {
                                 .put("startTime", Instant.now()) // getStartTime())
                                 .put("status", TransferStatusType.PAUSED.name()) // getStatus())
 //                                .add("tenant_id should not update")
-                                .put("totalSize", Math.pow(2, 20)) // getTotalSize())
-                                .put("transferRate", 9999) // getTransferRate())
+                                .put("totalSize", (long)Math.pow(2, 20)) // getTotalSize())
+                                .put("transferRate", 9999.0) // getTransferRate())
 //                                .addNull() // getParentTaskId())
 //                                .addNull() // getRootTaskId())
 //                                .add("this should not change") // getUuid())
-                                .put("totalFiles", 10) // getTotalFiles())
-                                .put("totalSkippedFiles", 9); // getTotalSkippedFiles());
+                                .put("totalFiles", 10L) // getTotalFiles())
+                                .put("totalSkippedFiles", 9L); // getTotalSkippedFiles());
 
                         TransferTask updatedTransferTask = new TransferTask(updatedJsonObject);
